show bench result for each timestep

+++ client/benchmarker.rb
@@ -9,7 +9,7 @@ require_relative './migrate'
 require_relative './worker'
 
 record_num = 100
-interval = 10
+interval = 300
 
 table_names = [[:users_id, :users_rating_secondary, :items_id, :items_quantity],
                [:users_id, :users_rating_secondary, :items_id,:items_quantity],
@@ -23,10 +23,10 @@ data[:items] = my_client.query("SELECT * FROM items LIMIT #{record_num}");
 cassandraManager = CassandraManager.new('rubis')
 
 tables = Tables.new.tables
-users_id = {query: 'SELECT users.* FROM users WHERE users.id = ? -- 8', frequency: [1,2,3]}
-users_rating = {query: 'SELECT users.* FROM users WHERE users.rating = ? -- 8', frequency: [1,2,3]}
-items_id = {query: 'SELECT items.* FROM items WHERE items.id=? -- 13', frequency: [1,2,3]}
-items_quantity = {query: 'SELECT items.* FROM items WHERE items.quantity=? -- 13', frequency: [1,2,3]}
+users_id = {query: 'SELECT users.* FROM users WHERE users.id = ? -- 8', frequency: [10000,20000,30000]}
+users_rating = {query: 'SELECT users.* FROM users WHERE users.rating = ? -- 8', frequency: [10000,20000,30000]}
+items_id = {query: 'SELECT items.* FROM items WHERE items.id=? -- 13', frequency: [30000,20000,10000]}
+items_quantity = {query: 'SELECT items.* FROM items WHERE items.quantity=? -- 13', frequency: [30000,20000,10000]}
 
 p_users_id = [
     [:users_id],
@@ -73,11 +73,13 @@ items_quantity_result = []
 
 cassandraManager.clearKeyspace
 table_names.each_with_index do |t_names, timestep|
-  cfs = t_names.map{|n| tables[n]}
-  cfs.each do |cf|
-    cassandraManager.dropTable(cf)
-    cassandraManager.createTable(cf)
-    cassandraManager.insert_data(data, cf)
+  if timestep == 0
+    cfs = t_names.map{|n| tables[n]}
+    cfs.each do |cf|
+      cassandraManager.dropTable(cf)
+      cassandraManager.createTable(cf)
+      cassandraManager.insert_data(data, cf)
+    end
   end
 
   initial[:users] = cassandraManager.executeQuery("SELECT rating, id FROM #{tables[:users_id].name}")
@@ -96,15 +98,15 @@ table_names.each_with_index do |t_names, timestep|
   }
   threads.each(&:join)
 
-  # x.report("users_id"){ users_id_result << tdp_users_id.plans[timestep].execute(initial[:users], cassandraManager)}
-  # x.report("users_rating"){ users_rating_result << tdp_users_rating.plans[timestep].execute(initial[:users], cassandraManager)}
-  # x.report("items_id"){ items_id_result << tdp_items_id.plans[timestep].execute(initial[:items], cassandraManager)}
-  # x.report("items_quantity"){ items_quantity_result << tdp_items_quantity.plans[timestep].execute(initial[:items], cassandraManager)}
-
-  sleep(interval * 2)
+  sleep(interval)
   workers.each(&:stop)
 
-  hoge = workers.map(&:read_from_child)
+  result = workers.map(&:read_from_child).map{|r| {query: r[:query], exec_time: r[:exec_time].sum / r[:exec_time].length}}
+
+  puts "=#{timestep}================"
+  result.each do |r|
+    puts "query: #{r[:query]}, exec_time: #{r[:exec_time]}"
+  end
 
   migration_plans.flatten(1).select{|mp| mp.start_timestep == timestep}.each do |mp|
     plans_for_timestep = [tdp_users_id, tdp_users_rating, tdp_items_id, tdp_items_quantity].map{|tdp| tdp.plans[timestep]}
